5758 - Rename helpers function
