+++ src/test/e2e/tests/06-validations/helpers/tables.ts
